Update Parser class to handle todo
